requirements.txt
customtkinter>=5.2.0
pyautogui>=0.9.54
pynput>=1.7.6
pillow>=9.5.0
pyinstaller>=5.13.0
selenium>=4.15.0 